Remove unecessary complexity and fixes  (#47)
* Remove unecessary complexity and fix #38

This offloads a lot of WTabHolder's work to WPanel by having each tab store its widgets in a WPanel. This removes a lot of convolution by only requiring interactions with one object and removes some code that would be duplicated in WPanel. It also solves #38 by allowing access to the body panel directly, with no need for offsetting.

* Fix weird tab behavior

As tabs weren't set as active/not active while being added, the last tab would always end up visible. Now when tabs are added, if it's tab 1, it's set as active, otherwise it's set as not active.